% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_quarto.R
\name{create_template_yaml}
\alias{create_template_yaml}
\title{Create Template YAML Configuration}
\usage{
create_template_yaml(template_name, logo_path = NULL)
}
\description{
Create Template YAML Configuration
}
\keyword{internal}
