% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_control.R
\name{dvc_commit}
\alias{dvc_commit}
\title{Commit Changes to DVC}
\usage{
dvc_commit(path, message)
}
\arguments{
\item{path}{Character vector of file paths to commit}

\item{message}{Commit message}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Commit Changes to DVC
}
