% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_control.R
\name{dvc_push}
\alias{dvc_push}
\title{Push Data to DVC Remote}
\usage{
dvc_push(path = NULL, remote = NULL)
}
\arguments{
\item{path}{Optional character vector of specific paths to push}

\item{remote}{Optional name of the remote to push to}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Push Data to DVC Remote
}
