% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvc_tracking.R
\name{dvc_track}
\alias{dvc_track}
\title{Track files with DVC after writing}
\usage{
dvc_track(path, message = NULL)
}
\arguments{
\item{path}{The path to the file that was written}

\item{message}{An optional commit message for DVC}
}
\value{
The input path (invisibly) to allow for further piping
}
\description{
This function adds DVC tracking to files that have been written using tidyverse
write functions. It is designed to be used in a pipe chain after write operations.
}
\examples{
\dontrun{
data |>
  readr::write_csv("data/processed/mydata.csv") |>
  dvc_track("Updated processed data")
}
}
