% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_cli_error}
\alias{format_cli_error}
\title{Format command-line errors into user-friendly messages}
\usage{
format_cli_error(cmd, output, status, default_msg = NULL)
}
\arguments{
\item{cmd}{The command that was executed}

\item{output}{The output from the command}

\item{status}{The exit status of the command}

\item{default_msg}{Default message to show if output is empty}
}
\value{
A formatted error message
}
\description{
Format command-line errors into user-friendly messages
}
\keyword{internal}
