% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_cmd_output}
\alias{format_cmd_output}
\title{Format command output for display}
\usage{
format_cmd_output(output, success = TRUE, cmd = NULL)
}
\arguments{
\item{output}{The command output to format}

\item{success}{Whether the command was successful}

\item{cmd}{The command that was run}
}
\value{
Formatted output suitable for display
}
\description{
Format command output for display
}
\keyword{internal}
