% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_control.R
\name{git_clone}
\alias{git_clone}
\title{Clone a Git Repository}
\usage{
git_clone(url, path, branch = NULL)
}
\arguments{
\item{url}{URL of the repository to clone}

\item{path}{Path where to clone the repository}

\item{branch}{Name of the branch to clone. Default is NULL (clones default branch).}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Clones a Git repository into a new directory.
}
\examples{
\dontrun{
git_clone("https://github.com/user/repo.git", "my_analysis")
git_clone("https://github.com/user/repo.git", "my_analysis", branch = "develop")
}
}
