% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_control.R
\name{git_init_repo}
\alias{git_init_repo}
\title{Initialize a Git Repository}
\usage{
git_init_repo()
}
\value{
Invisibly returns TRUE if successful
}
\description{
Initializes a new Git repository in the current directory.
}
\examples{
\dontrun{
git_init_repo()
}
}
