% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvc_tracking.R
\name{write_rds_dvc}
\alias{write_rds_dvc}
\title{Write RDS with DVC tracking}
\usage{
write_rds_dvc(object, file, message = NULL, ...)
}
\arguments{
\item{object}{Object to save}

\item{file}{Path to write to}

\item{message}{Optional DVC commit message}

\item{...}{Additional arguments passed to saveRDS}
}
\value{
The input object (invisibly) to allow for further piping
}
\description{
A wrapper around saveRDS that automatically tracks the output file with DVC
}
\examples{
\dontrun{
model |> write_rds_dvc("models/model.rds", "Updated model")
}
}
